+++ packages/fiori/src/ShellBar.ts
@@ -3,12 +3,10 @@ import { renderFinished } from "@ui5/webcomponents-base/dist/Render.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
-import event from "@ui5/webcomponents-base/dist/decorators/event.js";
+import event from "@ui5/webcomponents-base/dist/decorators/event-strict.js";
 import i18n from "@ui5/webcomponents-base/dist/decorators/i18n.js";
 import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
 import ResizeHandler from "@ui5/webcomponents-base/dist/delegate/ResizeHandler.js";
-import type AriaRole from "@ui5/webcomponents-base/dist/types/AriaRole.js";
-import AriaHasPopup from "@ui5/webcomponents-base/dist/types/AriaHasPopup.js";
 import { isSpace, isEnter } from "@ui5/webcomponents-base/dist/Keys.js";
 import ListItemStandard from "@ui5/webcomponents/dist/ListItemStandard.js";
 import List from "@ui5/webcomponents/dist/List.js";
@@ -26,7 +24,12 @@ import "@ui5/webcomponents-icons/dist/bell.js";
 import "@ui5/webcomponents-icons/dist/overflow.js";
 import "@ui5/webcomponents-icons/dist/grid.js";
 import "@ui5/webcomponents-icons/dist/slim-arrow-down.js";
-import type { Timeout, ClassMap, AccessibilityAttributes } from "@ui5/webcomponents-base/dist/types.js";
+import type {
+	Timeout,
+	ClassMap,
+	AccessibilityAttributes,
+	AriaRole,
+} from "@ui5/webcomponents-base";
 import type ListItemBase from "@ui5/webcomponents/dist/ListItemBase.js";
 import type PopoverHorizontalAlign from "@ui5/webcomponents/dist/types/PopoverHorizontalAlign.js";
 import type ShellBarItem from "./ShellBarItem.js";
@@ -49,9 +52,8 @@ import {
 	SHELLBAR_OVERFLOW,
 } from "./generated/i18n/i18n-defaults.js";
 
-type LowercaseString<T> = T extends string ? Lowercase<T> : never;
 type ShellBarLogoAccessibilityAttributes = {
-	role?: Extract<LowercaseString<AriaRole>, "button" | "link">,
+	role?: Extract<AriaRole, "button" | "link">,
 	name?: string,
 }
 type ShellBarProfileAccessibilityAttributes = Pick<AccessibilityAttributes, "name" | "expanded" | "hasPopup">;
@@ -165,13 +167,7 @@ const HANDLE_RESIZE_DEBOUNCE_RATE = 200; // ms
  * @param {HTMLElement} targetRef dom ref of the activated element
  * @public
  */
-@event<ShellBarNotificationsClickEventDetail>("notifications-click", {
-	detail: {
-		/**
-		 * @public
-		 */
-		targetRef: { type: HTMLElement },
-	},
+@event("notifications-click", {
 	cancelable: true,
 	bubbles: true,
 })
@@ -181,13 +177,7 @@ const HANDLE_RESIZE_DEBOUNCE_RATE = 200; // ms
  * @param {HTMLElement} targetRef dom ref of the activated element
  * @public
  */
-@event<ShellBarProfileClickEventDetail>("profile-click", {
-	detail: {
-		/**
-		 * @public
-		 */
-		targetRef: { type: HTMLElement },
-	},
+@event("profile-click", {
 	bubbles: true,
 })
 
@@ -198,13 +188,7 @@ const HANDLE_RESIZE_DEBOUNCE_RATE = 200; // ms
  * @param {HTMLElement} targetRef dom ref of the activated element
  * @public
  */
-@event<ShellBarProductSwitchClickEventDetail>("product-switch-click", {
-	detail: {
-		/**
-		 * @public
-		 */
-		targetRef: { type: HTMLElement },
-	},
+@event("product-switch-click", {
 	cancelable: true,
 	bubbles: true,
 })
@@ -215,13 +199,7 @@ const HANDLE_RESIZE_DEBOUNCE_RATE = 200; // ms
  * @since 0.10
  * @public
  */
-@event<ShellBarLogoClickEventDetail>("logo-click", {
-	detail: {
-		/**
-		 * @public
-		 */
-		targetRef: { type: HTMLElement },
-	},
+@event("logo-click", {
 	bubbles: true,
 })
 
@@ -233,13 +211,7 @@ const HANDLE_RESIZE_DEBOUNCE_RATE = 200; // ms
  * @since 0.10
  * @public
  */
-@event<ShellBarMenuItemClickEventDetail>("menu-item-click", {
-	detail: {
-		/**
-		 * @public
-		 */
-		item: { type: HTMLElement },
-	},
+@event("menu-item-click", {
 	bubbles: true,
 	cancelable: true,
 })
@@ -253,16 +225,20 @@ const HANDLE_RESIZE_DEBOUNCE_RATE = 200; // ms
  * @public
  */
 
-@event<ShellBarSearchButtonEventDetail>("search-button-click", {
-	detail: {
-		targetRef: { type: HTMLElement },
-		searchFieldVisible: { type: Boolean },
-	},
+@event("search-button-click", {
 	cancelable: true,
 	bubbles: true,
 })
 
 class ShellBar extends UI5Element {
+	eventDetails!: {
+		"notifications-click": ShellBarNotificationsClickEventDetail,
+		"profile-click": ShellBarProfileClickEventDetail,
+		"product-switch-click": ShellBarProductSwitchClickEventDetail,
+		"logo-click": ShellBarLogoClickEventDetail,
+		"menu-item-click": ShellBarMenuItemClickEventDetail,
+		"search-button-click": ShellBarSearchButtonEventDetail,
+	}
 	/**
 	 * Defines the `primaryTitle`.
 	 *
@@ -535,7 +511,7 @@ class ShellBar extends UI5Element {
 	}
 
 	_menuItemPress(e: CustomEvent<ListSelectionChangeEventDetail>) {
-		const shouldContinue = this.fireDecoratorEvent<ShellBarMenuItemClickEventDetail>("menu-item-click", {
+		const shouldContinue = this.fireDecoratorEvent("menu-item-click", {
 			item: e.detail.selectedItems[0],
 		});
 		if (shouldContinue) {
@@ -544,8 +520,8 @@ class ShellBar extends UI5Element {
 	}
 
 	_logoPress() {
-		this.fireDecoratorEvent<ShellBarLogoClickEventDetail>("logo-click", {
-			targetRef: this.shadowRoot!.querySelector(".ui5-shellbar-logo")!,
+		this.fireDecoratorEvent("logo-click", {
+			targetRef: this.shadowRoot!.querySelector<HTMLElement>(".ui5-shellbar-logo")!,
 		});
 	}
 
@@ -739,7 +715,7 @@ class ShellBar extends UI5Element {
 
 	_handleSearchIconPress() {
 		const searchButtonRef = this.shadowRoot!.querySelector<Button>(".ui5-shellbar-search-button")!;
-		const defaultPrevented = !this.fireDecoratorEvent<ShellBarSearchButtonEventDetail>("search-button-click", {
+		const defaultPrevented = !this.fireDecoratorEvent("search-button-click", {
 			targetRef: searchButtonRef,
 			searchFieldVisible: this.showSearchField,
 		});
@@ -802,13 +778,13 @@ class ShellBar extends UI5Element {
 		const notificationIconRef = this.shadowRoot!.querySelector<Button>(".ui5-shellbar-bell-button")!,
 			target = e.target as HTMLElement;
 
-		this._defaultItemPressPrevented = !this.fireDecoratorEvent<ShellBarNotificationsClickEventDetail>("notifications-click", {
+		this._defaultItemPressPrevented = !this.fireDecoratorEvent("notifications-click", {
 			targetRef: notificationIconRef.classList.contains("ui5-shellbar-hidden-button") ? target : notificationIconRef,
 		});
 	}
 
 	_handleProfilePress() {
-		this.fireDecoratorEvent<ShellBarProfileClickEventDetail>("profile-click", {
+		this.fireDecoratorEvent("profile-click", {
 			targetRef: this.shadowRoot!.querySelector<Button>(".ui5-shellbar-image-button")!,
 		});
 	}
@@ -821,7 +797,7 @@ class ShellBar extends UI5Element {
 		const buttonRef = this.shadowRoot!.querySelector<Button>(".ui5-shellbar-button-product-switch")!,
 			target = e.target as HTMLElement;
 
-		this._defaultItemPressPrevented = !this.fireDecoratorEvent<ShellBarProductSwitchClickEventDetail>("product-switch-click", {
+		this._defaultItemPressPrevented = !this.fireDecoratorEvent("product-switch-click", {
 			targetRef: buttonRef.classList.contains("ui5-shellbar-hidden-button") ? target : buttonRef,
 		});
 	}
@@ -1201,7 +1177,7 @@ class ShellBar extends UI5Element {
 			overflow: {
 				"title": this._overflowText,
 				"accessibilityAttributes": {
-					hasPopup: this.accessibilityAttributes.overflow?.hasPopup || AriaHasPopup.Menu.toLowerCase(),
+					hasPopup: this.accessibilityAttributes.overflow?.hasPopup || "menu" as const,
 					expanded: overflowExpanded === undefined ? this._overflowPopoverExpanded : overflowExpanded,
 				},
 			},