+++ packages/main/src/FormGroup.ts
@@ -86,16 +86,12 @@ class FormGroup extends UI5Element implements IFormItem {
 	@property()
 	itemSpacing: `${FormItemSpacing}` = "Normal";
 
-	@property()
-	labelSpan = "S12 M4 L4 XL4";
-
 	onBeforeRendering() {
 		this.processFormItems();
 	}
 
 	processFormItems() {
 		this.items.forEach((item: FormItem) => {
-			item.labelSpan = this.labelSpan;
 			item.itemSpacing = this.itemSpacing;
 		});
 	}