+++ packages/main/src/InputPopover.d.ts
@@ -1,3 +1,0 @@
-import type Input from "./Input";
-
-export default Input;