+++ packages/main/src/TabContainer.ts
@@ -1,10 +1,10 @@
 import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
-import type { AccessibilityAttributes, StyleData } from "@ui5/webcomponents-base/dist/types.js";
+import type { AccessibilityAttributes } from "@ui5/webcomponents-base/dist/types.js";
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
-import event from "@ui5/webcomponents-base/dist/decorators/event.js";
+import event from "@ui5/webcomponents-base/dist/decorators/event-strict.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
-import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
+import jsxRenderer from "@ui5/webcomponents-base/dist/renderer/JsxRenderer.js";
 import ResizeHandler from "@ui5/webcomponents-base/dist/delegate/ResizeHandler.js";
 import { renderFinished } from "@ui5/webcomponents-base/dist/Render.js";
 import slideDown from "@ui5/webcomponents-base/dist/animations/slideDown.js";
@@ -32,6 +32,8 @@ import arraysAreEqual from "@ui5/webcomponents-base/dist/util/arraysAreEqual.js"
 import { findClosestPosition, findClosestPositionsByKey } from "@ui5/webcomponents-base/dist/util/dragAndDrop/findClosestPosition.js";
 import Orientation from "@ui5/webcomponents-base/dist/types/Orientation.js";
 import DragRegistry from "@ui5/webcomponents-base/dist/util/dragAndDrop/DragRegistry.js";
+import handleDragOver from "@ui5/webcomponents-base/dist/util/dragAndDrop/handleDragOver.js";
+import handleDrop from "@ui5/webcomponents-base/dist/util/dragAndDrop/handleDrop.js";
 import type { SetDraggedElementFunction } from "@ui5/webcomponents-base/dist/util/dragAndDrop/DragRegistry.js";
 import longDragOverHandler from "@ui5/webcomponents-base/dist/util/dragAndDrop/longDragOverHandler.js";
 import MovePlacement from "@ui5/webcomponents-base/dist/types/MovePlacement.js";
@@ -61,7 +63,7 @@ import OverflowMode from "./types/OverflowMode.js";
 import type { IButton } from "./Button.js";
 
 // Templates
-import TabContainerTemplate from "./generated/templates/TabContainerTemplate.lit.js";
+import TabContainerTemplate from "./TabContainerTemplate.js";
 
 // Styles
 import tabContainerCss from "./generated/themes/TabContainer.css.js";
@@ -69,7 +71,7 @@ import ResponsivePopoverCommonCss from "./generated/themes/ResponsivePopoverComm
 
 type TabContainerPopoverOwner = "start-overflow" | "end-overflow" | TabInStrip;
 
-const tabStyles: Array<StyleData> = [];
+const tabStyles: Array<string> = [];
 const PAGE_UP_DOWN_SIZE = 5;
 
 type TabContainerStripInfo = {
@@ -165,7 +167,7 @@ interface ITab extends UI5Element {
 		tabContainerCss,
 		ResponsivePopoverCommonCss,
 	],
-	renderer: litRender,
+	renderer: jsxRenderer,
 	template: TabContainerTemplate,
 	dependencies: [
 		Button,
@@ -183,17 +185,7 @@ interface ITab extends UI5Element {
  * @public
  * @since 2.0.0
  */
-@event<TabContainerTabSelectEventDetail>("tab-select", {
-	detail: {
-		/**
-		 * @public
-		 */
-		tab: { type: HTMLElement },
-		/**
-		 * @public
-		 */
-		tabIndex: { type: Number },
-	},
+@event("tab-select", {
 	bubbles: true,
 	cancelable: true,
 })
@@ -206,17 +198,7 @@ interface ITab extends UI5Element {
  * @public
  * @since 2.0.0
  */
-@event<TabContainerMoveEventDetail>("move-over", {
-	detail: {
-		/**
-		 * @public
-		 */
-		source: { type: Object },
-		/**
-		 * @public
-		 */
-		destination: { type: Object },
-	},
+@event("move-over", {
 	bubbles: true,
 	cancelable: true,
 })
@@ -228,20 +210,15 @@ interface ITab extends UI5Element {
  * @param {object} destination Contains information about the destination of the moved element. Has `element` and `placement` properties.
  * @public
  */
-@event<TabContainerMoveEventDetail>("move", {
-	detail: {
-		/**
-		 * @public
-		 */
-		source: { type: Object },
-		/**
-		 * @public
-		 */
-		destination: { type: Object },
-	},
+@event("move", {
 	bubbles: true,
 })
 class TabContainer extends UI5Element {
+	eventDetails!: {
+		"tab-select": TabContainerTabSelectEventDetail;
+		"move-over": TabContainerMoveEventDetail;
+		"move": TabContainerMoveEventDetail;
+	}
 	/**
 	 * Defines whether the tab content is collapsed.
 	 * @default false
@@ -377,7 +354,7 @@ class TabContainer extends UI5Element {
 	_handleResizeBound: () => void;
 	_setDraggedElement?: SetDraggedElementFunction;
 
-	static registerTabStyles(styles: StyleData) {
+	static registerTabStyles(styles: string) {
 		tabStyles.push(styles);
 	}
 
@@ -530,7 +507,7 @@ class TabContainer extends UI5Element {
 	}
 
 	@longDragOverHandler("[data-ui5-stable=overflow-start],[data-ui5-stable=overflow-end],[role=tab]")
-	_onHeaderDragOver(e: DragEvent, isLongDragOver: boolean) {
+	_onHeaderDragOver(e: DragEvent, isLongDragOver?: boolean) {
 		if (!(e.target instanceof HTMLElement) || !e.target.closest("[data-ui5-stable=overflow-start],[data-ui5-stable=overflow-end],[role=tab],[role=separator]")) {
 			this.dropIndicatorDOM!.targetReference = null;
 			return;
@@ -550,36 +527,17 @@ class TabContainer extends UI5Element {
 			e.preventDefault();
 		} else if (closestPosition) {
 			const dropTarget = (closestPosition.element as TabInStrip).realTabReference;
-			let placements = closestPosition.placements;
 
 			if (dropTarget === draggedElement) {
-				placements = placements.filter(placement => placement !== MovePlacement.On);
+				closestPosition.placements = closestPosition.placements.filter(placement => placement !== MovePlacement.On);
 			}
 
-			const acceptedPlacement = placements.find(placement => {
-				const dragOverPrevented = !this.fireDecoratorEvent<TabContainerMoveEventDetail>("move-over", {
-					source: {
-						element: draggedElement!,
-					},
-					destination: {
-						element: dropTarget,
-						placement,
-					},
-				});
-
-				if (dragOverPrevented) {
-					e.preventDefault();
-					this.dropIndicatorDOM!.targetReference = closestPosition.element;
-					this.dropIndicatorDOM!.placement = placement;
-					return true;
-				}
-
-				return false;
-			});
-
-			if (acceptedPlacement === MovePlacement.On && (closestPosition.element as TabInStrip).realTabReference.items.length) {
+			const { targetReference, placement } = handleDragOver(e, this, closestPosition, dropTarget);
+			this.dropIndicatorDOM!.targetReference = targetReference;
+			this.dropIndicatorDOM!.placement = placement;
+			if (placement === MovePlacement.On && (closestPosition.element as TabInStrip).realTabReference.items.length) {
 				popoverTarget = closestPosition.element;
-			} else if (!acceptedPlacement) {
+			} else if (!placement) {
 				this.dropIndicatorDOM!.targetReference = null;
 			}
 		}
@@ -596,21 +554,8 @@ class TabContainer extends UI5Element {
 			return;
 		}
 
-		e.preventDefault();
-		const draggedElement = DragRegistry.getDraggedElement()!;
-
-		this.fireDecoratorEvent<TabContainerMoveEventDetail>("move", {
-			source: {
-				element: draggedElement,
-			},
-			destination: {
-				element: (this.dropIndicatorDOM!.targetReference as TabInStrip).realTabReference,
-				placement: this.dropIndicatorDOM!.placement,
-			},
-		});
-
+		handleDrop(e, this, (this.dropIndicatorDOM!.targetReference as TabInStrip).realTabReference, this.dropIndicatorDOM!.placement);
 		this.dropIndicatorDOM!.targetReference = null;
-		draggedElement.focus();
 	}
 
 	_moveHeaderItem(tab: Tab, e: KeyboardEvent) {
@@ -640,7 +585,7 @@ class TabContainer extends UI5Element {
 		});
 
 		const acceptedPosition = positions.find(({ element, placement }) => {
-			return !this.fireDecoratorEvent<TabContainerMoveEventDetail>("move-over", {
+			return !this.fireDecoratorEvent("move-over", {
 				source: {
 					element: tab,
 				},
@@ -652,7 +597,7 @@ class TabContainer extends UI5Element {
 		});
 
 		if (acceptedPosition) {
-			this.fireDecoratorEvent<TabContainerMoveEventDetail>("move", {
+			this.fireDecoratorEvent("move", {
 				source: {
 					element: tab,
 				},
@@ -707,7 +652,7 @@ class TabContainer extends UI5Element {
 			return;
 		}
 
-		const placementAccepted = !this.fireDecoratorEvent<TabContainerMoveEventDetail>("move-over", {
+		const placementAccepted = !this.fireDecoratorEvent("move-over", {
 			source: {
 				element: draggedElement,
 			},
@@ -751,7 +696,7 @@ class TabContainer extends UI5Element {
 
 		e.preventDefault();
 
-		this.fireDecoratorEvent<TabContainerMoveEventDetail>("move", {
+		this.fireDecoratorEvent("move", {
 			source: {
 				element: draggedElement,
 			},
@@ -970,7 +915,7 @@ class TabContainer extends UI5Element {
 	 * @returns true if the tab selection is successful, false if it was prevented
 	 */
 	selectTab(selectedTab: Tab, selectedTabIndex: number) {
-		if (!this.fireDecoratorEvent<TabContainerTabSelectEventDetail>("tab-select", { tab: selectedTab, tabIndex: selectedTabIndex })) {
+		if (!this.fireDecoratorEvent("tab-select", { tab: selectedTab, tabIndex: selectedTabIndex })) {
 			return false;
 		}