+++ packages/main/src/ToolbarSelect.ts
@@ -1,15 +1,15 @@
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
-import event from "@ui5/webcomponents-base/dist/decorators/event.js";
+import event from "@ui5/webcomponents-base/dist/decorators/event-strict.js";
 import type ValueState from "@ui5/webcomponents-base/dist/types/ValueState.js";
 
 import { registerToolbarItem } from "./ToolbarRegistry.js";
 
 // Templates
 
-import ToolbarSelectTemplate from "./generated/templates/ToolbarSelectTemplate.lit.js";
-import ToolbarPopoverSelectTemplate from "./generated/templates/ToolbarPopoverSelectTemplate.lit.js";
+import ToolbarSelectTemplate from "./ToolbarSelectTemplate.js";
+import ToolbarPopoverSelectTemplate from "./ToolbarPopoverSelectTemplate.js";
 import ToolbarItem from "./ToolbarItem.js";
 import Select from "./Select.js";
 import Option from "./Option.js";
@@ -45,13 +45,7 @@ type ToolbarSelectChangeEventDetail = SelectChangeEventDetail;
  * @param {HTMLElement} selectedOption the selected option.
  * @public
  */
-@event<ToolbarSelectChangeEventDetail>("change", {
-	detail: {
-		/**
-		* @public
-		*/
-		selectedOption: { type: HTMLElement },
-	},
+@event("change", {
 	bubbles: true,
 	cancelable: true,
 })
@@ -70,6 +64,11 @@ type ToolbarSelectChangeEventDetail = SelectChangeEventDetail;
  */
 @event("close")
 class ToolbarSelect extends ToolbarItem {
+	eventDetails!: ToolbarItem["eventDetails"] & {
+		change: ToolbarSelectChangeEventDetail;
+		open: void;
+		close: void;
+	}
 	/**
 	 * Defines the width of the select.
 	 *